move scheduler into its own package
